fist basic UI
Not a UI really, just a way to test things in the interface.
The plugin adds a new "Index" in the simple search selector on the
OPAC.
On the selector you can choose "Search with AI".
Then if you type in a request in natural language and click on submit
button, you will get the reply from Mistral in the JS console.log +
alert.
This is very basic and needs a lot of improvement but it is a first
stone :)